+++ docs/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 
-## Upcoming release
+## 2.1.1
+Released 2 March 2025.
 
 * Fixed clone listeners being called on value types, if they were cloned while boxed into `object`s.
 * Fixed `HashSetCloneEngine` not working when not using the `HashSet<>` type directly, but for example through the `ISet<>` interface.